Week 2 assignment- To Do List Component completed
